+++ core/rpc/src/rpc_impl/build_tx.rs
@@ -35,10 +35,6 @@ use std::iter::FromIterator;
 use std::str::FromStr;
 use std::vec;
 
-const fn ckb(num: u64) -> u64 {
-    num * BYTE_SHANNONS
-}
-
 #[derive(Default, Clone, Debug)]
 pub struct CellWithData {
     pub cell: packed::CellOutput,
@@ -55,131 +51,7 @@ impl CellWithData {
 }
 
 impl<C: CkbRpc> MercuryRpcImpl<C> {
-    pub(crate) async fn inner_build_adjust_account_transaction(
-        &self,
-        payload: AdjustAccountPayload,
-    ) -> InnerResult<Option<TransactionCompletionResponse>> {
-        if payload.asset_info.asset_type == AssetType::CKB {
-            return Err(RpcErrorMessage::AdjustAccountOnCkb);
-        }
-
-        let account_number = payload.account_number.clone().unwrap_or(1) as usize;
-        let extra_ckb = payload.extra_ckb.clone().unwrap_or_else(|| ckb(1));
-        let fee_rate = payload.fee_rate.unwrap_or(1000);
-        let item: Item = payload.item.clone().try_into()?;
-        let from = parse_from(payload.from.clone())?;
-        let mut estimate_fee = ckb(1);
-
-        loop {
-            let res = self
-                .adjust_account_with_fixed_fee(
-                    account_number,
-                    payload.asset_info.clone(),
-                    from.clone(),
-                    item.clone(),
-                    extra_ckb,
-                    estimate_fee,
-                )
-                .await?;
-
-            if res.is_none() {
-                return Ok(None);
-            }
-
-            let res = res.unwrap();
-            if res.2 {
-                return Ok(Some(res.0));
-            }
-
-            let tx_size = calculate_tx_size_with_witness_placeholder(
-                res.clone().0.tx_view,
-                res.clone().0.signature_entries,
-            ) as u64;
-
-            let mut actual_fee = fee_rate.saturating_mul(tx_size) / 1000;
-            if actual_fee * 1000 < fee_rate.saturating_mul(tx_size) {
-                actual_fee += 1;
-            }
-
-            if estimate_fee < actual_fee {
-                estimate_fee += BYTE_SHANNONS;
-                continue;
-            } else {
-                let tx_view = self.update_tx_view_change_cell(
-                    res.clone().0.tx_view,
-                    Address::new(self.network_type, res.clone().1),
-                    estimate_fee,
-                    actual_fee,
-                )?;
-                let adjust_response =
-                    TransactionCompletionResponse::new(tx_view, res.clone().0.signature_entries);
-                return Ok(Some(adjust_response));
-            }
-        }
-    }
-
-    pub(crate) async fn adjust_account_with_fixed_fee(
-        &self,
-        account_number: usize,
-        asset_info: AssetInfo,
-        from: Vec<Item>,
-        item: Item,
-        extra_ckb: u64,
-        fee: u64,
-    ) -> InnerResult<Option<(TransactionCompletionResponse, AddressPayload, bool)>> {
-        let mut asset_set = HashSet::new();
-        asset_set.insert(asset_info.clone());
-        let live_acps = self
-            .get_live_cells_by_item(
-                item.clone(),
-                asset_set,
-                None,
-                None,
-                Some((**ACP_CODE_HASH.load()).clone()),
-                None,
-            )
-            .await?;
-        let live_acps_len = live_acps.len();
-
-        if live_acps_len == account_number {
-            return Ok(None);
-        }
-
-        let sudt_type_script = self.build_sudt_type_script(asset_info.udt_hash.0.to_vec());
-
-        if live_acps_len < account_number {
-            let res = self
-                .build_create_acp_transaction_fixed_fee(
-                    from,
-                    account_number - live_acps_len,
-                    sudt_type_script,
-                    item,
-                    extra_ckb,
-                    fee,
-                )
-                .await?;
-            Ok(Some((
-                TransactionCompletionResponse::new(res.0.into(), res.1),
-                res.2,
-                res.3,
-            )))
-        } else {
-            let res = self
-                .build_collect_asset_fixed_fee(
-                    live_acps,
-                    live_acps_len - account_number,
-                    extra_ckb,
-                    fee,
-                )
-                .await?;
-
-            Ok(Some((
-                TransactionCompletionResponse::new(res.0.into(), res.1),
-                res.2,
-                res.3,
-            )))
-        }
-    }
+    
 
     pub(crate) async fn inner_build_deposit_transaction(
         &self,
@@ -802,6 +674,7 @@ impl<C: CkbRpc> MercuryRpcImpl<C> {
                 amount_required: required_udt as i128,
             }],
             Some(payload.from.source.clone()),
+            &mut 0,
             &mut inputs_part_3,
             &mut script_set,
             &mut signature_entries,
@@ -920,110 +793,9 @@ impl<C: CkbRpc> MercuryRpcImpl<C> {
         cells_data.push(data);
 
         Ok(())
-    }
+    }    
 
-    async fn build_create_acp_transaction_fixed_fee(
-        &self,
-        from: Vec<Item>,
-        acp_need_count: usize,
-        sudt_type_script: packed::Script,
-        item: Item,
-        extra_ckb: u64,
-        fee: u64,
-    ) -> InnerResult<(TransactionView, Vec<SignatureEntry>, AddressPayload, bool)> {
-        let mut ckb_needs = fee + MIN_CKB_CAPACITY + extra_ckb;
-        let mut outputs_data: Vec<packed::Bytes> = Vec::new();
-        let mut outputs = Vec::new();
-        let mut change_address = None;
-
-        for i in 0..acp_need_count {
-            let capacity = STANDARD_SUDT_CAPACITY + ckb(extra_ckb);
-            let output_cell = self.build_acp_cell(
-                Some(sudt_type_script.clone()),
-                self.get_secp_lock_hash_by_item(item.clone())?.0.to_vec(),
-                capacity,
-            );
-
-            if i == 0 {
-                change_address = Some(AddressPayload::from_script(
-                    &output_cell.lock(),
-                    self.network_type,
-                ));
-            }
-
-            outputs.push(output_cell);
-            outputs_data.push(Bytes::new().pack());
-            ckb_needs += capacity;
-        }
-
-        let mut inputs = Vec::new();
-        let mut input_capacity_sum = 0u64;
-        let mut script_set = HashSet::new();
-        let mut signature_entries = HashMap::new();
-
-        if from.is_empty() {
-            self.pool_live_cells_by_items(
-                vec![item],
-                ckb_needs,
-                vec![],
-                None,
-                &mut input_capacity_sum,
-                &mut inputs,
-                &mut script_set,
-                &mut signature_entries,
-            )
-            .await?;
-        } else {
-            self.pool_live_cells_by_items(
-                from,
-                ckb_needs,
-                vec![],
-                None,
-                &mut input_capacity_sum,
-                &mut inputs,
-                &mut script_set,
-                &mut signature_entries,
-            )
-            .await?;
-        }
-
-        script_set.insert(SECP256K1.to_string());
-        script_set.insert(SUDT.to_string());
-
-        let tx_view = self.build_transaction_view(&inputs, outputs, outputs_data, script_set, 0);
-
-        let mut sigs_entry = signature_entries
-            .into_iter()
-            .map(|(_k, v)| v)
-            .collect::<Vec<_>>();
-        sigs_entry.sort();
-
-        Ok((tx_view, sigs_entry, change_address.unwrap(), false))
-    }
-
-    fn build_acp_cell(
-        &self,
-        type_script: Option<packed::Script>,
-        lock_args: Vec<u8>,
-        capacity: u64,
-    ) -> packed::CellOutput {
-        let lock_script = self
-            .builtin_scripts
-            .get(ACP)
-            .cloned()
-            .expect("Impossible: get built in script fail")
-            .script
-            .as_builder()
-            .args(lock_args.pack())
-            .build();
-        packed::CellOutputBuilder::default()
-            .type_(type_script.pack())
-            .lock(lock_script)
-            .capacity(capacity.pack())
-            .build()
-    }
-
-    fn build_sudt_type_script(&self, type_args: Vec<u8>) -> packed::Script {
+    pub(crate) fn build_sudt_type_script(&self, type_args: Vec<u8>) -> packed::Script {
         self.builtin_scripts
             .get(SUDT)
             .cloned()
@@ -1332,9 +1104,10 @@ impl<C: CkbRpc> MercuryRpcImpl<C> {
     ) -> InnerResult<()> {
         self.pool_live_cells_by_items(
             items.to_owned(),
-            required_ckb as i64,
+            required_ckb,
             vec![],
             source,
+            &mut 0,
             inputs,
             script_set,
             signature_entries,
@@ -1392,6 +1165,7 @@ impl<C: CkbRpc> MercuryRpcImpl<C> {
                 amount_required: required_udt as i128,
             }],
             source,
+            &mut 0,
             inputs,
             script_set,
             signature_entries,
@@ -1454,15 +1228,6 @@ fn get_pool_capacity(inputs: &[DetailedCell]) -> InnerResult<u64> {
         })
         .sum();
     Ok(pool_capacity)
-}
-
-fn parse_from(from_set: HashSet<JsonItem>) -> InnerResult<Vec<Item>> {
-    let mut ret: Vec<Item> = Vec::new();
-    for ji in from_set.into_iter() {
-        ret.push(ji.try_into()?);
-    }
-
-    Ok(ret)
 }
 
 pub fn calculate_tx_size_with_witness_placeholder(
